Ease bar rank continuously, add ffmpeg WinGet fallback
Bars now glide through fractional rank positions between periods
instead of the renderer re-sorting by instantaneous value every
substep and snapping to a new slot the instant an overtake happens.

Also ports app.py's ffmpeg-resolution fallback (glob the WinGet
install path when shutil.which misses) into video_encoder.py, so a
long-running dev server doesn't need a restart after ffmpeg is
installed.

+++ bar-race-studio/backend/app/models/render_job.py
@@ -10,7 +10,7 @@ class FrameRow(BaseModel):
     category: str | None
     image_url: str | None
     value: float
-    rank: int
+    rank: float
 
 
 class FramesResponse(BaseModel):
